Refactoring of pretty printer.
