+++ backends/ze/gen_ze_library.rb
@@ -4,77 +4,6 @@ def ZE_BIT(i)
   1 << i
 end
 
-def popcount(x)
-  raise "Unsigned integer needed!" if x < 0
-  c = 0
-  while x != 0
-    c += 1 if x & 1 != 0
-    x >>= 1
-  end
-  c
-end
-
-def firstbitpos(x)
-  raise "Signed positive integer needed!" if x <= 0
-  r = 0
-  while x & 1 == 0
-    r += 1
-    x >>= 1
-  end
-  r
-end
-
-def print_bitfield(name, enum)
-  special_values = []
-  members = []
-  default = nil
-  counter = 0
-  enum.members.each { |m|
-    if m.val
-      counter = m.val.kind_of?(String) ? eval(m.val) : m.val
-    else
-      counter += 1
-    end
-    if counter > 0 && popcount(counter) == 1
-      members.push [ m.name.to_sym, firstbitpos(counter) ]
-    else
-      if counter == 0
-        default = m.name.to_sym
-      else
-        r = [ m.name.to_sym ]
-        if m.val
-          r.push m.val
-        else
-          r.push counter
-        end
-        special_values.push r
-      end
-    end
-  }
-  print_lambda = lambda { |m|
-    "#{m[0].inspect}, #{m[1]}"
-  }
-  puts <<EOF
-  #{to_class_name(name)} = zebitmask #{to_ffi_name(name)},
-    [ #{members.collect(&print_lambda).join(",\n      ")} ]
-EOF
-  if default
-    puts <<EOF
-  #{to_class_name(name)}.default = #{default.inspect}
-EOF
-  end
-  if !special_values.empty?
-    puts <<EOF
-  # #{special_values.collect(&print_lambda).join(",\n  # ")}
-EOF
-  end
-  if to_ffi_name(name).match("_flag_t")
-    puts "  #{to_class_name(name)}s = #{to_class_name(name)}"
-    puts "  typedef #{to_ffi_name(name)}, #{to_ffi_name(name).gsub("_flag_t", "_flags_t")}"
-  end
-  puts "\n"
-end
-
 def print_version_enum_struct(name)
   puts <<EOF
   class #{to_class_name(name)} < FFI::ZEStruct
@@ -96,107 +25,20 @@ end
 
 def print_enum(name, enum)
   return print_version_enum_struct(name) if name.match(/_version_t\z/)
-  members = enum.members.collect { |m|
-    r = [ m.name.to_sym ]
-    r.push m.val if m.val
-    r
-  }
-  print_lambda = lambda { |m|
-    s = "#{m[0].inspect}"
-    s << ", #{m[1]}" if m.size == 2
-    s
-  }
-  puts <<EOF
-  #{to_class_name(name)} = zeenum #{to_ffi_name(name)},
-    [ #{members.collect(&print_lambda).join(",\n      ")} ]
-
-EOF
+  if enum.members.find { |m| m.val && m.val.kind_of?(String) && m.val.match("ZE_BIT") } || enum.members.all? { |m| m.name.include?("_FLAG_") }
+    print_bitfield_with_namespace(:ZE, name, enum, check_flags: true)
+  else
+    print_enum_with_namespace(:ZE, name, enum)
+  end
 end
 
 def print_ze_object(object)
-  puts <<EOF
-  typedef :pointer, #{to_ffi_name(object)}
-
-EOF
+  print_object(object)
 end
+
+print_ffi_module(:ZE, enclosing_module:false)
 
 puts <<EOF
-require 'ffi'
-module FFI
-
-  class ZEStruct < Struct
-    def to_h
-      members.zip(values).to_h
-    end
-
-    def to_s
-      s = "{ "
-      s << to_h.each.collect { |k, v|
-        if v.kind_of? Array
-          "\#{k}: [ \#{v.join(", ")} ]"
-        else
-          "\#{k}: \#{v}"
-        end
-      }.join(", ")
-      s << " }"
-      s
-    end
-  end
-  class ZEEnum < Enum
-  end
-  class ZEUnion < Union
-  end
-  class ZEBitmask < Bitmask
-    def default=(default)
-      @default = default
-    end
-
-    def []
-      if @default && query.size == 1
-        if query[0] == 0
-          return @default
-        elsif query[0] == @default
-          return 0
-        end
-      end
-      super
-    end
-
-    def to_native(query, ctx)
-      return 0 if query.nil?
-      flat_query = [query].flatten
-      return 0 if flat_query.size == 1 && query[0] == @default
-      super
-    end
-
-    def from_native(val, ctx)
-      return [@default] if val == 0 && @default
-      super
-    end
-  end
-  module Library
-    def zebitmask(*args)
-      generic_enum(FFI::ZEBitmask, *args)
-    end
-    def zeenum(*args)
-      generic_enum(FFI::ZEEnum, *args)
-    end
-  end
-
-  class Struct::InlineArray
-    def to_s
-      s = "[ "
-      s << to_a.join(", ")
-      s << " ]"
-    end
-  end
-
-  class Pointer
-    def to_s
-      "0x%016x" % address
-    end
-  end
-end
 module ZE
   ZE_MAX_IPC_HANDLE_SIZE = 64
   ZE_MAX_UUID_SIZE = 16
@@ -357,59 +199,23 @@ module ZE
 EOF
 
 def print_union(name, union)
-  members = union.to_ffi
-  print_lambda = lambda { |m|
-    s = "#{m[0]}, "
-    if m[1].kind_of?(Array)
-      s << "[ #{m[1][0]}, #{m[1][1]} ]"
-    else
-      s << "#{m[1]}"
-    end
-    s
-  }
-  puts <<EOF
-  class #{to_class_name(name)} < FFI::ZEUnion
-    layout #{members.collect(&print_lambda).join(",\n"+" "*11)}
-  end
-  typedef #{to_class_name(name)}.by_value, #{to_ffi_name(name)}
-
-EOF
+  print_union_with_namespace(:ZE, name, union)
 end
 
 def print_struct(name, struct)
-  members = struct.to_ffi
-  print_lambda = lambda { |m|
-    s = "#{m[0]}, "
-    if m[1].kind_of?(Array)
-      s << "[ #{m[1][0]}, #{m[1][1]} ]"
-    else
-      s << "#{m[1]}"
-    end
-    s
-  }
-  puts <<EOF
-  class #{to_class_name(name)} < FFI::ZEStruct
-EOF
+  prepends = []
   if to_class_name(name).match("UUID")
     if to_class_name(name).match("ZEKernelUUID")
-  puts <<EOF
-    prepend KUUID
-EOF
+      prepends << "KUUID"
     else
-  puts <<EOF
-    prepend UUID
-EOF
+      prepends << "UUID"
     end
   elsif to_class_name(name).match(/Handle\z/)
-puts <<EOF
-    prepend Handle
-EOF
+      prepends << "Handle"
   end
-  puts <<EOF
-    layout #{members.collect(&print_lambda).join(",\n"+" "*11)}
-EOF
-  if members.first[0] == ":stype" && !$struct_type_reject.include?(name)
-    puts <<EOF
+
+  if struct.to_ffi.first[0] == ":stype" && !$struct_type_reject.include?(name)
+    initializer = <<EOF
 
     def initialize(*args)
       super(*args)
@@ -429,27 +235,17 @@ EOF
         raise "Unrecognized namespace for #{ename}"
       end.upcase
     ename = $struct_type_conversion_table[ename] if $struct_type_conversion_table[ename]
-    puts <<EOF
+  
+    initializer << <<EOF
         self[:stype] = :#{ename}
       end
     end
 EOF
+  else
+    initializer = nil
   end
-  puts <<EOF
-  end
-  typedef #{to_class_name(name)}.by_value, #{to_ffi_name(name)}
 
-EOF
-end
-
-def print_function_pointer_type(name, func)
-  type, params = func.to_ffi
-  puts <<EOF
-  callback #{to_ffi_name(name)},
-           [ #{params.join(",\n"+" "*13)} ],
-           #{type}
-
-EOF
+  print_struct_with_namespace(:ZE, name, struct, prepends: prepends, initializer: initializer, close: false)
 end
 
 $int_scalars.each { |k, v|
@@ -463,11 +259,7 @@ EOF
 $all_types.each { |t|
   if t.type.kind_of? YAMLCAst::Enum
     enum = $all_enums.find { |e| t.type.name == e.name }
-    if enum.members.find { |m| m.val && m.val.kind_of?(String) && m.val.match("ZE_BIT") } || enum.members.all? { |m| m.name.include?("_FLAG_") }
-      print_bitfield(t.name, enum)
-    else
-      print_enum(t.name, enum)
-    end
+    print_enum(t.name, enum)
   elsif $objects.include?(t.name)
     print_ze_object(t.name)
   elsif t.type.kind_of? YAMLCAst::Struct
